fix: ajusta responsividade e organização de elemntos para entrega; fix(delete user): proibe exclusão de root user

+++ frontend/src/pages/perfil/index.jsx
@@ -10,15 +10,16 @@ import perfil from "./perfil.module.scss"
 import { toast } from 'react-toastify';
 import "react-toastify/dist/ReactToastify.css";
 import api from '../../api'
-
 import { useContext } from "react";
 import { UserContext } from "../../components/Context/userContext";
+import { useConfirmation } from "../../components/ModalConfirmation/handleHook";
 
 const Perfil = () => {
     const navigate = useNavigate();
+    const {confirm, ConfirmationModal} = useConfirmation()
+    const {logout} = useContext(UserContext)
     const [load, setLoad] = useState(false);
     const [lock, setLock] = useState(false)
-    //const [text, setText] = useState('')
     const [name, setName] = useState('')
     
     const [email, setEmail] = useState('')
@@ -36,8 +37,7 @@ const Perfil = () => {
         try{
             let req = await api.delete('/user')
          
-            
-            if(req.status == 200){
+            if(req.status == 204){
                 
                 setLoad(false)
                 setTimeout(() => {
@@ -65,50 +65,54 @@ const Perfil = () => {
     }
     }
 
-    async function updateUser(){
-        //let tokenId = localStorage.getItem("id")
-        const userData = {name: name, email: email}
-        
+    async function updateUser() {
+        const userData = {};
+        if (name !== client) userData.name = name;
+        if (email !== mail) userData.password = email;
 
-        try{
-            let req = await api.patch('/user', userData)
-            if(req.status == 200){
-                
-                setLoad(false)
-                 setTimeout(() => {
-                                    setLoad(false)
-                                    toast.success('Usuário atualizado!');
-                                }, 1000);
-                                
-            }
+        if (Object.keys(userData).length === 0) {
+            toast.info("Nenhuma informação foi alterada.");
+            setLoad(false)
+            return;
         }
-        catch (error) {
-            console.log(error)
-            if(error.response){
-                    setLoad(false)
-                    setTimeout(() => {
-                        toast.error(error.response.data.message);
-                    }, 1000);
-            
-            }else{
-                setLoad(false)
-                setTimeout(() => {
-                toast.error('Servidor não respondeu. Verifique sua conexão ou tente mais tarde.');
+
+        try {
+            const req = await api.patch('/user', userData);
+            if (req.status === 200) {
+            setLoad(false);
+            setTimeout(() => {
+                toast.success('Usuário atualizado!');
             }, 1000);
             }
-    }
+        } catch (error) {
+            console.log(error);
+            setLoad(false);
+            setTimeout(() => {
+            if (error.response) {
+                toast.error(error.response.data.message);
+            } else {
+                toast.error('Servidor não respondeu. Verifique sua conexão ou tente mais tarde.');
+            }
+            }, 1000);
+        }
     }
 
+
     
-    const loading = (func) => {
-        setLoad(true);
+    const loading = async (func) => {
         if(func == 'update'){
-            updateUser()
+            if(await confirm("atualizar seus dados")){
+                setLoad(true);
+                updateUser()
+                location.reload()
+            }
         }
-        
         else if(func == 'deleteUser'){
-            
-            deleteUser()
+            if(await confirm("excluir sua conta")){
+                setLoad(true);
+                deleteUser()
+                logout()
+            }
         }
         
     }
@@ -153,8 +157,8 @@ const Perfil = () => {
                 <form className={perfil.formProfile}>
                     <label>Nome</label>
                     <InputFormEdit onChange={(event) => handleChange(event, setName)} place={name} disable={lock} type="text"></InputFormEdit>
-                    <label>E-mail</label>
-                    <InputFormEdit onChange={(event) => handleChange(event, setEmail)}  place={mail} disable={lock} type="email"></InputFormEdit>
+                    <label>Senha</label>
+                    <InputFormEdit onChange={(event) => handleChange(event, setEmail)}  place={mail} disable={lock} type="password"></InputFormEdit>
 
                     <label>Pontos</label>
                     <div className={perfil.divPoints}>
@@ -168,11 +172,12 @@ const Perfil = () => {
                     <div className={perfil.formButtons}>
                         <button className={perfil.buttonCancelar} disabled={lock} onClick={() => navigate("/")}>Cancelar</button>
                         <button type="button" className={perfil.buttonAlterar} disabled={lock} onClick={() => loading('update')}>Alterar</button>
-                        <button className={perfil.buttonExcluirConta} disabled={lock} onClick={() => loading('deleteUser')}>Excluir Conta</button>
+                        <button type="button" className={perfil.buttonExcluirConta} disabled={lock} onClick={() => loading('deleteUser')}>Excluir Conta</button>
                     </div>
                 </form>
             </div>
             <Footer/>
+            <ConfirmationModal></ConfirmationModal>
         </>
     );
 };
